requirements.txt
asgiref==3.5.0
black==22.1.0
canvasapi==2.2.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.4
commonmark==0.9.1
decorator==5.1.1
docutils==0.18.1
falcon==3.1.0
h11==0.13.0
humanize==4.0.0
idna==3.3
lockfile==0.12.2
mypy==0.941
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.5.1
Pygments==2.12.0
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2021.3
requests==2.27.1
rich==12.4.4
six==1.16.0
tomli==2.0.1
types-python-dateutil==2.8.10
typing_extensions==4.1.1
urllib3==1.26.8
uvicorn==0.17.6
validators==0.18.2
